using puppet to install puppet-lint

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,5 +1,6 @@
 # kill the process
 exec { 'killmenow':
+  creates => 'killmenow',
   path    => 'alx-system_engineering-devops/0x0A-configuration_management',
   creates => 'alx-system_engineering-devops/0x0A-configuration_management',
   command => 'pkill -9 -f killmenow'
